Move edit-template into /edit folder
